Pridanie delete tlacitka

+++ frontend/src/components/Box/BoxProfil/BoxProfil.js
@@ -5,9 +5,22 @@ import axios from 'axios';
 import { useState } from "react";
 
 const BoxProfil = (props) => {
-   const [id, setId] = useState(props.id);
- 
-   const buttonDelete = async () => {
+//useState ukladanie hodnot.. id, heslo z id a input heslo
+  const [id, setId] = useState(props.id); 
+  const [dataPass, setDataPass] = useState("");
+  const [inputPassword, setinputPassword]= useState("");
+  //Zistenie podla api endpoint heslo podla id
+    fetch("https://endpoint.vybavenka.sk/api/users")
+    .then(zistit=>zistit.json())
+    .then(data=>data.filter(potvrdenie=>potvrdenie.id===props.id))
+    .then(dataa=>console.log(setDataPass(dataa[0].deletePassword)));
+
+
+    //Funkcia na spustenie mazania inzeratu.. a kontrola.. hesla
+   const buttonDelete = async (event) => {
+    event.preventDefault();
+    //Zistovanie hesla
+    if(dataPass===inputPassword){
      try {
        console.log(id);
        await axios.delete(`https://endpoint.vybavenka.sk/api/users/${id}`, {
@@ -21,7 +34,7 @@ const BoxProfil = (props) => {
        console.log(error);
      }
    };
- 
+  }
    return (
      <div className="boxProfil">
        {/* Zobrazenie v Menu inzeratu Profil */}
@@ -32,11 +45,20 @@ const BoxProfil = (props) => {
        <div className="boxFullProfil">
          <BoxObsahProfil prezivkaUser={props.prezivkaUser} userEmail={props.userEmail} />
        </div>
+       <form onSubmit={buttonDelete}>     
+         <div className="boxProfilDeleteFull">
+           <h4 className="nadpisDeleteHeslo">Vymazat inzerat</h4>
+       <div className="divdeleteInput">      
+        <input className="deleteInput" type="password" placeholder="Vaše Heslo"  onChange={(event) => setinputPassword(event.target.value)}/>
+        </div>
        {/* Vymazanie inzeratu v Profile */}
        <div className="boxDeleteProfil">
-         <button className="boxDeleteButton" onClick={buttonDelete}>Vymazať Inzerat</button>
+         <button className="boxDeleteButton" type="submit">Vymazať</button>
        </div>
      </div>
+     </form>
+     </div>
+
    );
  };
